basic maths problems

+++ reverse_digits.cpp
@@ -4,7 +4,7 @@ using namespace std;
 void reverseDigits() {
     int n;
     cin >> n;
-    int reverseNum = 0;
+    long reverseNum = 0;
     while (n > 0) {
         int lastDigit = n % 10;
         reverseNum = reverseNum * 10 + lastDigit;
